Refactor UserUpdated to use PreviousAttributes class
